backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
certifi==2025.1.31
    # via
    #   httpcore
    #   httpx
cffi==1.17.1
    # via cryptography
click==8.1.8
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   tqdm
    #   uvicorn
cryptography==44.0.2
    # via pyjwt
distro==1.9.0
    # via openai
dnspython==2.7.0
    # via email-validator
docstring-parser==0.16
    # via llmeasytools
email-validator==2.2.0
    # via fastapi
fastapi[standard]==0.115.12
    # via -r requirements.in
fastapi-cli[standard]==0.0.7
    # via fastapi
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.7
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   openai
idna==3.10
    # via
    #   anyio
    #   email-validator
    #   httpx
jinja2==3.1.6
    # via fastapi
jiter==0.9.0
    # via openai
llmeasytools==0.25
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
openai==1.72.0
    # via -r requirements.in
pycparser==2.22
    # via cffi
pydantic==2.11.3
    # via
    #   fastapi
    #   llmeasytools
    #   openai
pydantic-core==2.33.1
    # via pydantic
pygments==2.19.1
    # via rich
pyjwt[crypto]==2.10.1
    # via -r requirements.in
python-dotenv==1.1.0
    # via
    #   -r requirements.in
    #   uvicorn
python-multipart==0.0.20
    # via fastapi
pyyaml==6.0.2
    # via uvicorn
rich==14.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.14.1
    # via fastapi-cli
shellingham==1.5.4
    # via typer
sniffio==1.3.1
    # via
    #   anyio
    #   openai
starlette==0.46.1
    # via fastapi
tqdm==4.67.1
    # via openai
typer==0.15.2
    # via fastapi-cli
typing-extensions==4.13.2
    # via
    #   anyio
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   rich-toolkit
    #   typer
    #   typing-inspection
typing-inspection==0.4.0
    # via pydantic
uvicorn[standard]==0.34.0
    # via
    #   fastapi
    #   fastapi-cli
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
